+++ Projects/Core/Core/Sources/Extension/UIKit/UIColor+.swift
@@ -1,0 +1,25 @@
+//
+//  UIColor+.swift
+//  Core
+//
+//  Created by JDeoks on 7/12/25.
+//
+
+
+import UIKit
+
+extension UIColor {
+  convenience init(hex: String, alpha: CGFloat = 1.0) {
+    var hexSanitized = hex.trimmingCharacters(in: .whitespacesAndNewlines)
+    hexSanitized = hexSanitized.replacingOccurrences(of: "#", with: "")
+    
+    var rgb: UInt64 = 0
+    Scanner(string: hexSanitized).scanHexInt64(&rgb)
+    
+    let red = CGFloat((rgb & 0xFF0000) >> 16) / 255.0
+    let green = CGFloat((rgb & 0x00FF00) >> 8) / 255.0
+    let blue = CGFloat(rgb & 0x0000FF) / 255.0
+    
+    self.init(red: red, green: green, blue: blue, alpha: alpha)
+  }
+}
